test(tgi): refactor tests
